tasks/load-opa-properties/requirements.txt
google-cloud-bigquery==3.14.1
google-cloud-storage==2.*
functions-framework==3.5.0
